Create Lab03_Q3
Design a class named "Product" to encapsulate product details while utilizing access specifiers appropriately. A parameterized constructor that initializes productId, productName, price, and quantityInStock when an object is created.
a)	Create an instance of the "Product" class, representing a product with an ID of 101, name "Laptop," price $800, and quantity in stock 10.
b)	Access and display the name, price, and quantity of the product.
c)	Attempt to modify the product's ID from external code (outside the class). Observe and comment on the result.